requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
yfinance>=0.2.0
lightgbm>=4.0.0
